tests/wpt/tests/html/browsers/offline/application-cache-api/api_status_obsolete-manual.html

<!DOCTYPE HTML>
<html manifest="../resources/manifest/clock.manifest">
  <head>
    <title>Offline Application Cache - API_status_OBSOLETE</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <ol>
      <li>Remove the manifest file (manifest/clock.manifest) from the server.</li>
      <li>Refresh the page.</li>
    </ol>
    <div id="log"></div>

    <script>
      var t = async_test("obsolete status test"),
          cache = window.applicationCache;

      cache.onobsolete = t.step_func_done(function() {
        assert_equals(cache.status, cache.OBSOLETE, "cache.status should equals cache.OBSOLETE");
      });
    </script>
  </body>
</html>
